Comments and code cleanup for polymorphic
