+++ polymorphic/test_switch.ts
@@ -1,6 +1,6 @@
-import { PolymorphicSwitch, serializePolymorphicClass } from "./polymorphic.ts";
-import { Serializable } from "./serializable.ts";
-import { SerializeProperty } from "./serialize_property.ts";
+import { polymorphicClassFromJSON, PolymorphicSwitch } from "./polymorphic.ts";
+import { Serializable } from "../serializable.ts";
+import { SerializeProperty } from "../serialize_property.ts";
 
 abstract class MyBaseClass extends Serializable {
   @SerializeProperty()
@@ -31,12 +31,12 @@ class MyOtherPolyClass extends MyBaseClass {
   }
 }
 
-const myPolyClassInstance = serializePolymorphicClass(
+const myPolyClassInstance = polymorphicClassFromJSON(
   MyBaseClass,
   { "_class": "MyPolyClassImplementation", "parentValue": 11 },
 );
 
-const myOtherPolyClassInstance = serializePolymorphicClass(
+const myOtherPolyClassInstance = polymorphicClassFromJSON(
   MyBaseClass,
   { "_class": "MyOtherPolyClass", "someProperty": 0, "parentValue": 22 },
 );
@@ -52,7 +52,7 @@ console.log("=== Test full serialize ===");
 const data =
   `{"_class":"MyOtherPolyClass", "someProperty": 13, "parentValue": 33}`;
 
-const test = serializePolymorphicClass(MyBaseClass, data);
+const test = polymorphicClassFromJSON(MyBaseClass, data);
 
 console.log(test);
